added gps functionality, but not tested

+++ RaspberryPi/hardware/neo6m/neo6m.cpp
@@ -299,7 +299,7 @@ void Neo6M::disableNMEAMessage(const std::string &msg) {
     if(msg.length() != 3)
         throw std::length_error("NMEA message identifier must contain 3 characters");
     char buffer[24]; //don't forget \0
-    if(snprintf(buffer, sizeof(buffer), "PUBX,40,%s,0,0,0,0,0,0", msg.c_str()) != (sizeof(buffer) - 1))
+    if(snprintf(buffer, sizeof(buffer), "PUBX,40,%s,0,0,0,0,0,0", msg.c_str()) != (int) (sizeof(buffer) - 1))
         throw std::runtime_error("creating NMEA message: " + std::string(strerror(errno)));
     sendNMEAMessage(buffer);
 }
@@ -308,7 +308,7 @@ void Neo6M::enableNMEAMessage(const std::string &msg) {
     if(msg.length() != 3)
         throw std::length_error("NMEA message identifier must contain 3 characters");
     char buffer[24]; //don't forget \0
-    if(snprintf(buffer, sizeof(buffer), "PUBX,40,%s,0,1,0,0,0,0", msg.c_str()) != (sizeof(buffer) - 1))
+    if(snprintf(buffer, sizeof(buffer), "PUBX,40,%s,0,1,0,0,0,0", msg.c_str()) != (int) (sizeof(buffer) - 1))
         throw std::runtime_error("creating NMEA message: " + std::string(strerror(errno)));
     sendNMEAMessage(buffer);
 }
@@ -316,7 +316,7 @@ void Neo6M::enableNMEAMessage(const std::string &msg) {
 void Neo6M::sendNMEAMessage(const std::string &msg) {
     char buffer[msg.length() + 7]; //don't forget \0
     uint8_t checksum = calcNMEAChecksum(msg);
-    if(snprintf(buffer, sizeof(buffer), "$%s*%.2X\r\n", msg.c_str(), checksum) != (sizeof(buffer) - 1))
+    if(snprintf(buffer, sizeof(buffer), "$%s*%.2X\r\n", msg.c_str(), checksum) != (int) (sizeof(buffer) - 1))
         throw std::runtime_error("creating NMEA message: " + std::string(strerror(errno)));
     writeAll(buffer, sizeof(buffer)-1);
 }
